fix: refactored create property page and solved route protecting bug
